Created DAQ_windows branch to work on Windows specific version of the DAQ code as the python APIs are different on Windows and Linux
